+++ src/components/Employee/Deductions/index.ts
@@ -1,1 +1,1 @@
-export * from './Deductions'
+export * from './onboarding'
