+++ src/components/Employee/Deductions/onboarding/DeductionsForm/ChildSupportFormView.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from 'react-i18next'
 import type { Garnishment } from '@gusto/embedded-api-v-2025-11-15/models/components/garnishment'
 import type { PaymentPeriod } from '@gusto/embedded-api-v-2025-11-15/models/components/garnishmentchildsupport'
-import { useChildSupportGarnishmentForm } from '../shared/useChildSupportGarnishmentForm'
-import type { StateFieldEntry, CountyEntry } from '../shared/useChildSupportGarnishmentForm'
+import { useChildSupportGarnishmentForm } from '../../shared/useChildSupportGarnishmentForm'
+import type { StateFieldEntry, CountyEntry } from '../../shared/useChildSupportGarnishmentForm'
 import { Form } from '@/components/Common/Form'
 import { ActionsLayout } from '@/components/Common'
 import { Flex } from '@/components/Common/Flex/Flex'
