+++ src/components/Employee/Deductions/onboarding/DeductionsList/DeductionsList.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest'
 import { screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import type { Garnishment } from '@gusto/embedded-api-v-2025-11-15/models/components/garnishment'
-import type { UseDeductionsListReady } from '../shared/useDeductionsList'
+import type { UseDeductionsListReady } from '../../shared/useDeductionsList'
 import { DeductionsList } from './DeductionsList'
 import { renderWithProviders } from '@/test-utils/renderWithProviders'
