+++ src/components/Employee/Deductions/onboarding/DeductionsList/DeductionsList.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from 'react-i18next'
 import type { Garnishment } from '@gusto/embedded-api-v-2025-11-15/models/components/garnishment'
-import type { UseDeductionsListReady } from '../shared/useDeductionsList'
-import { formatDeductionAmount } from '../shared/formatDeductionAmount'
+import type { UseDeductionsListReady } from '../../shared/useDeductionsList'
+import { formatDeductionAmount } from '../../shared/formatDeductionAmount'
 import { useDataView, DataView } from '@/components/Common'
 import { ActionsLayout } from '@/components/Common'
 import { Flex } from '@/components/Common/Flex/Flex'
