+++ src/components/Employee/Deductions/onboarding/deductionsContextualComponents.tsx
@@ -1,8 +1,8 @@
 import { type Garnishment } from '@gusto/embedded-api-v-2025-11-15/models/components/garnishment'
+import { useDeductionsList } from '../shared/useDeductionsList'
 import { IncludeDeductions } from './IncludeDeductions/IncludeDeductions'
 import { DeductionsList } from './DeductionsList/DeductionsList'
 import { DeductionsForm } from './DeductionsForm/DeductionsForm'
-import { useDeductionsList } from './shared/useDeductionsList'
 import { BaseLayout } from '@/components/Base/Base'
 import { useFlow, type FlowContextInterface } from '@/components/Flow/useFlow'
 import { componentEvents } from '@/shared/constants'
