+++ src/components/Employee/Deductions/onboarding/index.ts
@@ -1,0 +1,1 @@
+export * from './Deductions'
